[v3-1-test] Remove global from breeze PROVIDER_DEPENDENCIES (#58857)
* Remove global from breeze PROVIDER_DEPENDENCIES

* Consolidate constants

* Refactor provider package tooling from global constants to provider dependencies

* Add typo annotation for mypy
(cherry picked from commit 9e6c15fd18bf14d24953070a9ba39228f8b8eb2a)
