+++ dev/breeze/src/airflow_breeze/utils/path_utils.py
@@ -271,6 +271,7 @@ BUILD_CACHE_PATH = AIRFLOW_ROOT_PATH / ".build"
 GENERATED_PATH = AIRFLOW_ROOT_PATH / "generated"
 CONSTRAINTS_CACHE_PATH = BUILD_CACHE_PATH / "constraints"
 PROVIDER_DEPENDENCIES_JSON_PATH = GENERATED_PATH / "provider_dependencies.json"
+PROVIDER_DEPENDENCIES_JSON_HASH_PATH = GENERATED_PATH / "provider_dependencies.json.sha256sum"
 PROVIDER_METADATA_JSON_PATH = GENERATED_PATH / "provider_metadata.json"
 UI_CACHE_PATH = BUILD_CACHE_PATH / "ui"
 AIRFLOW_TMP_PATH = AIRFLOW_ROOT_PATH / "tmp"
